Added option that allows usage of custom .mp3 files as alarm sound.
Added some code to make sure  app runs in the foreground so the alarm goes off.

Added some code that would let the app wake the device's screen if it is turned off when the alarm goes off.

Added some code that allows the app to alert the user even if the phone is locked.